+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webkit_constants.dart
@@ -131,4 +131,9 @@ class WebViewFeatureConstants {
   ///
   /// See https://developer.android.com/reference/androidx/webkit/WebViewFeature#WEB_AUTHENTICATION.
   static const String webAuthentication = 'WEB_AUTHENTICATION';
+
+  /// This feature covers [WebView.addDocumentStartJavaScript].
+  ///
+  /// See https://developer.android.com/reference/androidx/webkit/WebViewFeature#DOCUMENT_START_SCRIPT.
+  static const String documentStartScript = 'DOCUMENT_START_SCRIPT';
 }